requirements.txt
# Bot Framework Core
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-dialogs==4.15.0
botbuilder-ai==4.15.0
aiohttp>=3.9.0

# Azure Services (Optional - use if you have Azure Cognitive Services)
# azure-ai-textanalytics>=5.3.0  # Uncomment if using Azure Text Analytics
azure-identity>=1.15.0
applicationinsights>=0.11.10

# Database
pyodbc>=5.0.0
sqlalchemy>=2.0.0

# Machine Learning (Optional - install only if needed)
transformers>=4.36.0  # Uncomment for advanced NLP
torch>=2.1.0  # Uncomment if using PyTorch models
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Sentiment Analysis (lightweight)
vaderSentiment>=3.3.2

# Dashboard
streamlit>=1.29.0
plotly>=5.18.0

# Configuration
python-dotenv>=1.0.0

# Testing (Optional - for development)
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Logging
loguru>=0.7.0

# Utilities
requests>=2.31.0